+++ tests/testthat/test-igr_is_valid.R
@@ -8,7 +8,24 @@ test_that("basic invalid", {
   expect_equal(igr_is_valid("I"), FALSE)
   expect_equal(igr_is_valid(c("", "B1", "B126", "B12367", "B1234678", "B123456789", "B12345167890", "B123451678906")), rep(FALSE, 8))
   expect_equal(igr_is_valid(c("A1 234", "a123 4", "A1 2 34", "A 12 3 4", "A 123 4")), rep(FALSE, 5))
-  expect_equal(igr_is_valid(c("AA 1 2", "A1A2", "A12A")), rep(FALSE, 3))
+  expect_equal(igr_is_valid(c("AA 1 2", "A1A2", "A1234A")), rep(FALSE, 3))
+})
+
+test_that("tetrad true valid", {
+  expect_equal(igr_is_valid("A00A"), TRUE)
+  expect_equal(igr_is_valid(c("A00A", "Z00Z", "Z99Z", "N19N", "N19P", "N19Z")), rep(TRUE, 6))
+  expect_equal(igr_is_valid(c("A56Z", "A 56 Z", " A56Z", "A 56Z", "A 5 6 Z", "A 5 6Z", "A 56 Z ")), rep(TRUE, 7))
+})
+
+test_that("tetrad true invalid", {
+  expect_equal(igr_is_valid(c("AA", "A000A", "A0000A", "A000000A", "A00000000A", "A0000000000A")), rep(FALSE, 6))
+  expect_equal(igr_is_valid("A00O"), FALSE)
+})
+
+test_that("tetrad false", {
+  expect_equal(igr_is_valid("A00A", tetrad = FALSE), FALSE)
+  expect_equal(igr_is_valid(c("A00A", "Z00Z", "Z99Z", "N19N", "N19P", "N19Z"), tetrad = FALSE), rep(FALSE, 6))
+  expect_equal(igr_is_valid(c("A56Z", "A 56 Z", " A56Z", "A 56Z", "A 5 6 Z", "A 5 6Z", "A 56 Z "), tetrad = FALSE), rep(FALSE, 7))
 })
 
 test_that("edge invalid", {