+++ tests/testthat/test-st_irishgridrefs.R
@@ -11,12 +11,14 @@ test_that("basic generations", {
 test_that("all resolutions 100km > 1m", {
   expect_equal(st_irishgridrefs(x1_sf, precision = 100000), c("A"))
   expect_equal(st_irishgridrefs(x1_sf, precision = 10000), c("A00"))
+  expect_equal(st_irishgridrefs(x1_sf, precision = 2000), c("A00A"))
   expect_equal(st_irishgridrefs(x1_sf, precision = 1000), c("A0000"))
   expect_equal(st_irishgridrefs(x1_sf, precision = 100), c("A000000"))
   expect_equal(st_irishgridrefs(x1_sf, precision = 10), c("A00000000"))
   expect_equal(st_irishgridrefs(x1_sf, precision = 1), c("A0000000000"))
 
   expect_equal(st_irishgridrefs(x2_sf, precision = 100000), c("A", "Z"))
+  expect_equal(st_irishgridrefs(x2_sf, precision = 2000), c("A00A", "Z90A"))
   expect_equal(st_irishgridrefs(x2_sf, precision = 1), c("A0000000000", "Z9000000000"))
 })
 
@@ -31,7 +33,7 @@ test_that("only POINT geometry", {
 test_that("invalid precision detected", {
   expect_error(st_irishgridrefs(x1_sf, precision = 0), class = "unsupported_precision")
   expect_error(st_irishgridrefs(x1_sf, precision = 2), class = "unsupported_precision")
-  expect_error(st_irishgridrefs(x1_sf, precision = 2000), class = "unsupported_precision")
+  expect_error(st_irishgridrefs(x1_sf, precision = 5000), class = "unsupported_precision")
   expect_error(st_irishgridrefs(x1_sf, precision = "A"), class = "unsupported_precision")
 })