+++ NEWS.md
@@ -1,5 +1,7 @@
 # igr (development version)
 
+* Tetrad support added (2000 m precision, "DINTY" system, e.g. "N85H").
+
 # igr 0.1.1
 
 * Description extended and vignette build technique adjusted for CRAN compatibility.